pyproject.toml
[project]
name = "Websocket-Demo"
version = "0.1.0"
description = ""
requires-python = ">=3.12,<4.0"

dependencies = [
    "fastapi>=0.128.0",
    "opencv-python>=4.12.0.88",
    "pillow>=12.0.0",
    "python-multipart>=0.0.21",
    "uvicorn>=0.40.0",
    "websockets>=15.0.1",
]

[tool.uv]
dev-dependencies = [
    "pipdeptree==2.26.1",
]
